# PERF: use the libjoin fast paths in a few more cases