- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.3.0.rst` file if fixing a bug or adding a new feature.

Additional cases where the fast path can be used:
* left join with many:1 (right unique)
* right join with 1:many (left unique)

These are covered by existing tests. 

```
import pandas as pd
import numpy as np

left = pd.Index(np.arange(100_000)).repeat(10)
right = pd.Index(np.arange(100_000))

%timeit left.join(right, how="left")

# 22.7 ms ± 7.84 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
# 3.8 ms ± 60.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

```
import pandas as pd
import numpy as np

data = [f"i_{i:06}" for i in range(100_000)]
left = pd.Index(data, "string[pyarrow_numpy]")
right = pd.Index(data, "string[pyarrow_numpy]").repeat(10)

%timeit left.join(right, how="right")

# 393 ms ± 4.64 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# 150 ms ± 3.98 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```